pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "mcp-template"
version = "1.0.0"
description = "MCP Server Template using FastMCP"
readme = "README.md"
keywords = ["mcp", "llm", "ai", "claude", "model", "context", "protocol", "server"]
requires-python = ">=3.11"
dependencies = [
    "fastmcp>=2.3.3",
    "python-dotenv>=1.1.1",
    "openai>=2.4.0",
    "mcp>=1.1.2",
    "jsonschema>=4.25.1",
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Natural Language :: English",
    "Operating System :: OS Independent",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Typing :: Typed"
]

[project.scripts]
mcp-template = "main:main"

[tool.setuptools]
packages = ["src"]
py-modules = ["main"]

[tool.rye]
dev-dependencies = [
    "isort>=7.0.0",
    "ruff>=0.14.0",
    "pytest>=8.4.2",
]
